Debezium bridge not ignoring full initial updates
Debezium bridge is only forwarding data if it has been changed. One problem with this approach
is the initialization of the Debezium connected. When it is started, it is pushing the whole database state to
the stream. This is used to make sure that database and streaming system are in sync. The Debezium
bridge is now detecting this initialization state and forwarding the data even though it was not changed.