Split KafkaTopic CR creation from other CRs and allow makefile to removee kafkatopics before deploying flink
Strimzi is great. But one annoying behaviour is that it creates KafkaTopic CR for every topic it finds
on Kafka. This is in principle great, because KafkaTopic CRDs allow configuring all details of topics, like
replication. However, when an application creates a Kafka Topic, Strimzi creates a KafkaTopic CR which conflicts
with the CRs from the helm deployments. In order to be able to delete the KafkaTopics CR which are relevant, the
KafkaTopic CRs are created in separate files to be able to explicitly delete them before deploying the shacl job.

Closes #402